+++ pkg/apis/navigator/v1alpha1/types.go
@@ -26,7 +26,7 @@ type CassandraClusterSpec struct {
 	Sysctl    []string                   `json:"sysctl"`
 	NodePools []CassandraClusterNodePool `json:"nodePools"`
 	Image     CassandraImage             `json:"image"`
-	CqlPort   int64                      `json:"cqlPort"`
+	CqlPort   int32                      `json:"cqlPort"`
 }
 
 type CassandraImage struct {
